require node>=17 & relocate ts output
